+++ CHANGELOG.md
@@ -1,4 +1,15 @@
+## v0.3.0 (2015-10-12)
+
+* Enhancements
+  * Add last lines of supervisor log to system info pane. Only shows for superusers.
+  * Automatically update brando/villain static through brunch. 
+    See our updated `brunch-config.js`.
+
+* Bug fixes
+  * Fixed logrotate permissions bug in fabfile.
+
 ## v0.2.0 (2015-10-04)
+
 * Enhancements
   * Additional custom stylesheet option for admin. 
     SCSS for this is stored in `web/static/css/brando.custom.scss`, and compiled to `priv/static/css/brando.custom.css`.